Recolour light theme if system theme  clashes.
On some Linux the default theme has dark grey drop downs, which can't be recoloured (limitation in wxWidgets) and that looks bad with the built in light themes (Classic and Light).  So in these cases the theme's light colour and light buttons get recoloured.

Custom themes and dark theme don't get recoloured.