Fix posterisation.  Add blend themes option.
Posterisation was caused by having a bright theme cache and using a darker theme (in linux) so the light colours 'got stretched' and showed up the relatively few distinct values there.  So the Classic theme is now darker, and we expect (on Windows) that it will be recoloured to windows lighter colours.