Bug 1639 - Revert fix.
The fix was regarded as a workaround, and too draconian.  The problem is that Left and Right shortcuts are being captured (somehow) by our CommandManager when navigating in docked toolbars.  A full and proper fix rather than not binding these shortcuts would apply them to the toolbar in preference to the CommandManager, and so fix bug 1637, which Bug 1639 has now been marked as a duplicate of.