Bug 1637 - Keyboard interaction with many controls in toolbars broken
This progresses this bug.  Specifically on Windows, Left and Right arrows and Enter (non keypad version) should now no longer be a problem.  The fix is to not show these shortcuts in menus, because otherwise the menus will catch the events.  Unfortunately attempts to show these specific shortcuts in the menus bring back the bug.

Steps to reproduce for 1637 will need to be updated.