Add code for iota theme-bitmap
This #ifdeffed out TEST_CARD is useful for javascript code that shows how a
theme will look, when applied to Audacity.  With the code enabled, each
pixel of audacity is painted with a colour that encodes the location in
the theme image map that was used.

Net result - in javascript you can change any colour in the theme cache
and instantly see how Audacity will look.