Informative dialog, if action disallowed.
This commit is mostly about reducing duplicated code,
so that we can handle this in one place.

The disallowed dialog now can mention what action was
disallowed, the title can be more informative, and it may
mention what to do about it.

The commit also brings us a step closer to customisable toolbars,
thanks to some work on the EditToolbar.