Send Spectrum Settings help to its own page in the manual
This commit provides a different help landing page (Settings) for preferences that are used outside the normal Ctrl+P preferences dialog - for example, 'spectrum settings' accessed from the track menu.   This should work for any settings that are used on their own in that way, though spectrum settings is currently the only example.