Allow selected controls on SelectionToolbar
The toolbar can now show subsets of the start, center, length and end numeric controls.  The choice is made using a button with a pop-up menu.  The two most recently modified controls 'drive' the other two, whether they are visible or not.  If shown, the driven controls are indicated in the name above (also for screen readers).

Most users will opt to just show two of the controls, e.g. start and end.

The menu button (Classic theme) has three dots on it.  Other themes don't yet have the icon and will show as black.