Simplify logic of track control panel click to rearrange...
... Rearrange drag is now simply the default of all other hit tests.

The comment, as to the purpose of the removed call to TrackSelFunc, is
obsolete, because clicks on hidden buttons and sliders have recently been
disabled.

The mention of "the area left of the minimize button" is obsolete because there
is no such area any more after 2.2.0 layout changes.

Just before this commit, click in the minimize button but near the left edge
did not minimize, but started a track rearrange drag!  That should not be so.