Fix the test for removability of envelope points...
... Which went wrong in some examples of an effect applied to part of an
envelope that contains only one point; the correct result should have
a level 1 in the selection and unchanged levels left and right of it, no
matter whether the sole point was before, in, or after the selection.

Simplification may still remove that sole point yet still leave the levels
outside the selection correct.