Fix for Audacity crashing when running Jaws on Windows 10 1703
Problem: On Windows 10 1703, with the Jaws screen reader running, additional paint messages are sent to Audacity compared with when Jaws is not running. My assumption is this is probably a Jaws bug. In particular, when a project is closed, ToolDock::OnPaint, and AdornedRulerPanel::OnPaint are called.

Fix: changes ensure that these OnPaint functions can be called without causing a crash.