1678: Cursor to track start/end misbehaves when no tracks selected
Problem: With no tracks selected, the commands cursor to track start/end set the cursor position to a silly high value or undefined respectively.

Simple fix: Set the flags on the commands to TracksSelectedFlag, as they were before the menus were reorganised.

Note that after this fix, with tracks present, but none selected, these commands cause a disallowed dialog to open, saying that there was a problem with the last action. I don't know if this behaviour is now expected.