Fix PortMidi automake files and update PortMidi's m4
Several of these changes are needed to correctly access PortMidi's headers without hardcoding paths.  The change to the comment of USE_PORTMIDI is because that will actually be needed in a later commit.  The remainder just correct typos or other text that wasn't fixed from when I wrote the checklib based off of portsmf's.