Check for both USE_MIDI and USE_PORTMIDI for EXPERIMENTAL_MIDI_OUT
This allows building without portmidi.  In the future, if EXPERIMENTAL_MIDI_OUT is to become the default, the check for USE_PORTMIDI should be used instead; however, for now, only whether or not EXPERIMENTAL_MIDI_OUT is defined is controlled by it (which is functionally the same).

The one exception is for the library linking in AudacityApp; there it makes more sense to directly check USE_PORTMIDI.