Fix bug 1662 (crashing TCP menu commands on Windows) differently...
Previous fix was at commit f9865f39efe46224a1a77c742a789b2cc70f0c54

But then that had this strange effect on Mac:  when a menu command such as
rename brought up a modal dialog, then the first click in the dialog would pass
through to the tracks behind it, changing the selection, and only a repeated
click on the OK button would dismiss the dialog.

I don't know exactly why that happened.  This avoids it.