Fix more missing checks for USE_MIDI
This corrects the cases that were missed in #257, that don't affect compilation.

Some of these changes were from the track panel refactor where most of the note track UI didn't check for USE_MIDI.

The remainder are about using Open or drag+drop with a MIDI file.  The behavior is now to always import as an audio file if USE_MIDI is disabled (which means that FFMPEG's mod import will be used in that case), which seems reasonable.

This change also adds more information about the check for EXPERIMENTAL_MIDI_OUT that James Crook added in bcf21ad0fd (which was originally also included in this commit).

Finally, it replaces the incorrect HAVE_MIDI check with one for EXPERIMENTAL_MIDI_OUT.