Make exported label tracks locale-independent
Currently, when Audacity exports labels tracks numbers are formatted
with the "%f" specifier, which relies on the current user locale to pick
up the decimal separator.

This means that the numbers indicating the start and the end of the
labeled interval could end up using a comma as a decimal separator.

This makes the exported file less portable, especially in the case of
parsing the exported file with an external tool.

Audacity is already able to _import_ labels tracks independently of the
locale because it uses Internat::CompatibleToDouble(), so do something
symmetric at _export_ time, using Internat::ToString() to make the
exported data locale-independent (i.e. always using a dot as the decimal
separator).

Proposed in https://sourceforge.net/p/audacity/mailman/message/35534945/

NOTE:

When converting numbers to strings FLT_DIG is passed as the
digitsAfterDecimalPoint argument of Internat::ToString(), this is in
order to preserve the look and the alignment of the previous "%f" format
specifier; without that Internat::ToString() would strip trailing zeros.

Audacity requires C++11, this ensures that FLT_DIG is defined.