Possible fix for Mac library issue.
Fix off by one error in size.

Git-bisect indicated a problem (on Mac only) caused by a change in an image file.
It's possible that this is an error in code that we have 'got away with' until now.
We always have an extra zero on the end of the image data and included it in the count of the
size of stream to unpack.  For a particular image this may have triggered a memory corruption
problem.