Bug1590: Mac skips plugin cfg entries for .ny's of other versions...
... This conservative fix depends on where you put the executable bundle.

Assuing the executable is in /Applications, and the other Audacity version
(past or future) is there too (either bundled as for 2.1.3 and later, or
in a folder as for 2.1.2 and previous):

The .ny files that are part of the other version's shipment are ignored, and
therefore the Effects menu avoids duplicates.

But the .cfg file is not rewritten to eliminate them, in case you want to
switch back to the other version.  If that other version is 2.1.3 or earlier,
sorry but you will see duplicates.