Replace mentions of feedback@audacityteam.org with url of forum...
... There are two in AboutDialog.cpp, but these strings aren't actually used.

There are two in InconsistencyException.cpp, which ought never to be seen by
users, but should be treated as high priority bugs if they are.

There is a very old one in Sequence.cpp, which was never internationalized and
was only written to the log and likewise ought to be eliminated from happening
in pre-release testing.

Therefore, though this is a minor breach of string freeze, it's only the second
one above that users might ever see, and then only if a serious bug escaped.