Remove some unused strings that confused translators...
... "Light" was used in two different meanings, but one place was the obsolete
Leveller effect, whose code is removed now.  (I leave the files in place, to
make it easier for the curious to know it was once there, and find its history.)

i18n-hints added for the surviving use of "Light" to name a theme, and also for
some of the other themes.