Fix links to local help for buttons in Library preferences...
... There are four different code paths to test, for four different "Download"
buttons:

The two "Download" buttons in the Libraries preference dialog; and,
The Download button in each of the two dialog boxes that follow the two
"Locate..." buttons.

There are just two new help pages in a proper build of local help, each the
target of two buttons.

One of these pages is called "FAQ:Installing the FFmpeg-Import-Export Library"
rather than "FAQ:Installing the FFmpeg-Import/Export Library"

because there are some difficulties preventing links to pages with "/" in the
title, which I didn't try to fix.