Edit one link destination inside an internationalized string...
... Apply substituion to the result of translation lookup, assuming the
translator preserved the url verbatim.  (But in future, we should not have
urls inside translated strings.)

Problem is that the page faq_opening_and_saving_files.html will exist in the
2.2.0 manual online, but the anchor #midi will no longer exist on it.

This help link comes up only when import fails (perhaps the file is corrupt)
and the user presses the help button.

This help goes through an unusual code path that never uses local help.  That
could stand fixing in future.