Fix bug in pushing and popping sizers in notebook pages...
... Wrong sizer was pushed instead of the page's own, and was popped as
current sizer, when the first sub-sizer in the page was done.

But nothing was affected by this error, because notebook pages are used only
in the About dialog, and none of the pages have more than one directly nested
sizer.  So popping null as current sizer didn't matter because the stack did
not deepen again.

PushSizer() only after assigning mpSizer, as with other calls to it.