Completely replace PaUtil_GetTime with Pa_GetStreamTime
This requires some reordering of logic so that the PortAudio stream is not closed until we've finished using it for timing (although it is stopped) and equally is not used before it's started.  It also changes SystemTime to not be static (since it depends upon the stream).

Note that this basically still ends up calling PaUtil_GetTime for some OSes, just through the "intended" means.