Bug1784: Changed conditions for exceptions from Sequence...
... That for SetSamples was too strict, making needless errors in case of
harmless zero-length clips.

The one for Get was analogous to SetSamples.

That for Delete was too lax.  But reexamination shows the stricter condition
to be satisfied in all calls.  Sequence::Delete() can be reached only from
WaveClip::Clear and WaveClip::ClearAndAddCutLine(), and all calls to those
are in WaveTrack.cpp.