Update fr.po
Update the French translation.

I was on my way to update it to 100% but I noticed a reference to transifex. Since I'm not on transifex and I'm reluctant to be so and, before continuing to update the translation further, I'd like someone with commit access to the audacity git repo to confirm that it's OK for me to send a pull request with the full French GUI translation here on github.

Please advise. Thanks.