Fix subtle error revealed by the previous commit ...
... A "translated" effect family string, plus "/Enable",  was used as a
registry key for lookup!  It is wrong to key on a translation.

But in fact exhaustive search for "/Enable" shows that the only such strings
for which a written (by EffectsPrefs) registry value could be found were

AudioUnit
Ladspa
LV2
Nyquist
VAMP
VST

And none of these was actually a msgid in audacity.pot.  So nothing bad can
really have happened in other locale settings.