Clip VRuler black keys instead of hiding them
In my opinion it looks better to have them always visible (just shorter) at the top than to have them flicker in and out of existence while resizing tracks.  In the original form (where the track simply expanded when you dragged it) the hiding them behavior may have made sense, but now that we try to maintain the same number of notes, it does not.