Add support for zooming to show all notes
This is now the behavior of Zoom Reset, with Max Zoom being given the previous behavior.  Shift+Right-Click alternates between the two -- it zooms to all notes, unless it currently is zoomed to all notes in which case it performs a max zoom.

This also fixes Bug 1815, by performing the all-notes zoom when importing a track.