Disable part of the drop-out detection...
... The part that checks the previously unused statusFlags argument of
audacityAudioCallback can make zero-length labels.  I did provoke this into
happening repeatably on macOS using a debug build, zero buffer length in device
preferences, and a busy CPU running other programs, within just two minutes of
recording.

But close zooming in on the label, and listening, revealed nothing obviously
wrong, no click in the playback.  So I consider that a false positive.

But the part of the drop-out detection that would make nonzero length
labels, because the other AudioThread is lagging in its writes to disk --
this part remains.  Yet I have not yet provoked this into happening.