Update sv.po
Translated the strings lacking a translation and updated all fuzzy strings. I fear that there might be more strings in need to be updated that I've missed, since many of them wasn't marked as "fuzzy" when I opened this .po file in Poedit.

The reason why I closed my previous pull request is because I fixed a typo.