Protect Nyquist against effect-within-effect
The Effect code makes temporary copies of wavetracks.  If Nyquist invokes another effect,
its temporary copy is no longer valid.
My solution is to NOT apply the Nyquist effect results IF some other effect was successfully
invoked in creating them.  I do this by counting the effect invocations.  If it has changed, then
Nyquist output Is not applied back to the track.

The net result is that Nyquist prompt can invoke other effects, or process the audio itself,
but it cannot do both.